% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changepoint_screening.R
\name{changepoint_screening}
\alias{changepoint_screening}
\title{Changepoint multivariate Poisson time series, penality is L_{1,1}}
\usage{
changepoint_screening(fit, gamma)
}
\arguments{
\item{fit}{The output object from \code{changepoint_dp()}}

\item{gamma}{Tuning parameter to control the number of output changepoints. Cannot be \code{NA}}
}
\value{
a list with a vector \code{partition} denoting the resulting
partition and a vector \code{cusum_vec} denoting the resulting CUSUM at each of the middle indices
in \code{partition}
}
\description{
Does a screening based on the multidimensional CUSUM.
}
