% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis.R
\name{construct_AR_basis}
\alias{construct_AR_basis}
\title{Construction of the basis values for auto-regressive processes}
\usage{
construct_AR_basis(dat, thres_u = Inf, lag = 1)
}
\arguments{
\item{dat}{Count data, where each row represents a different time step and each
column represents a different variable}

\item{thres_u}{A positive threshold for the saturation effect}

\item{lag}{Amount of lag, should be a positive integer}
}
\value{
A matrix that has \code{nrow(dat)} rows and \code{ncol(dat)*lag}
columns.
}
\description{
The first few rows of the returned object will have 0's, implying that
the "initial values" of what is regressed on are 0.
}
