% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data.R
\name{create_data}
\alias{create_data}
\title{Create data}
\usage{
create_data(coef_list, partition, cov_type = "identity")
}
\arguments{
\item{coef_list}{list of numerics}

\item{partition}{vector}

\item{cov_type}{character}
}
\value{
list containing \code{X} and code{y}
}
\description{
Create data
}
