% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changepoint_dp.R
\name{.changepoint_dp}
\alias{.changepoint_dp}
\title{Title}
\usage{
.changepoint_dp(dat, thres_u = round(stats::quantile(dat[dat > 0], probs
  = 0.75)), lambda, gamma, min_spacing = 10, verbose = T)
}
\arguments{
\item{dat}{Count data, where each row represents a different time step and each
column represents a different variable}

\item{thres_u}{A positive threshold for the saturation effect}

\item{lambda}{Tuning parameter for the regression problem. Cannot be \code{NA}}

\item{gamma}{Tuning parameter to control the number of output changepoints. Cannot be \code{NA}}

\item{min_spacing}{Positive integer to dictate the smallest gap in changepoints. Needed for numerical stability
when calling \code{glmnet}}

\item{verbose}{Boolean}
}
\value{
A list containing \code{obj_val} (the objective value),
\code{partition} (the estimated partition) and \code{A_list} the estimated
adjecency matrices within each partition
}
\description{
This method ignore ALL trying changepoints of size less than \code{min_spacing}.
}
