% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stationary_timeseries.R
\name{generative_model}
\alias{generative_model}
\title{Generate a dataset from a true model}
\usage{
generative_model(nu, A, timesteps = 10, thres_u = 5,
  basis_function = construct_AR_basis, warning_val = 1000, ...)
}
\arguments{
\item{nu}{Vector in intercepts, one for each variable}

\item{A}{Square transition matrix}

\item{timesteps}{Number of time steps to proceed forward with}

\item{thres_u}{A positive threshold for the saturation effect}

\item{basis_function}{The function to generate the variables to regress onto, which
can possibly take addition inputs}

\item{warning_val}{If not \code{NA}, the code will stop if any of the
parameters of the Poisson exceed this number. The default is set to 1000.}

\item{...}{Addition parameters for the basis function}
}
\value{
A matrix that has \code{timesteps} rows and \code{ncol(A)} columns
of non-negative integers
}
\description{
Samples from the model in (2.1), (2.2) with saturation in (2.6) in
https://arxiv.org/pdf/1802.04838.pdf. The
construction of the basis can be passed in as a function.
}
