% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screening.R
\name{hausdorff}
\alias{hausdorff}
\title{Hausdorff distance}
\usage{
hausdorff(set1, set2, one.sided = FALSE)
}
\arguments{
\item{set1}{numeric}

\item{set2}{numeric}

\item{one.sided}{boolean}
}
\value{
numeric
}
\description{
Hausdorff distance
}
