% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_dim_buhlmann.R
\name{high_dim_buhlmann_estimate}
\alias{high_dim_buhlmann_estimate}
\title{High dimensional estimator - Buhlmann}
\usage{
high_dim_buhlmann_estimate(X, y, lambda, K = NA, gamma = NA,
  delta = 10, max_candidates = NA, verbose = F)
}
\arguments{
\item{X}{\code{n} by \code{d} matrix}

\item{y}{length \code{n} vector}

\item{lambda}{numeric}

\item{K}{numeric}

\item{gamma}{numeric}

\item{delta}{numeric}

\item{max_candidates}{numeric}

\item{verbose}{boolean}
}
\value{
list containing \code{partition} and \code{coef_list}
}
\description{
High dimensional estimator - Buhlmann
}
