% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_dim_feasible.R
\name{high_dim_feasible_estimate}
\alias{high_dim_feasible_estimate}
\title{High dimensional estimator (feasible)}
\usage{
high_dim_feasible_estimate(X, y, lambda, K = NA, tau = NA, M = 0,
  delta = 10, max_candidates = NA, verbose = F)
}
\arguments{
\item{X}{\code{n} by \code{d} matrix}

\item{y}{length \code{n} vector}

\item{lambda}{numeric}

\item{K}{numeric}

\item{tau}{numeric}

\item{M}{numeric}

\item{delta}{numeric}

\item{max_candidates}{numeric}

\item{verbose}{boolean}
}
\value{
list containing \code{partition} and \code{coef_list}
}
\description{
High dimensional estimator (feasible)
}
