% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_dim_infeasible.R
\name{high_dim_infeasible_estimate}
\alias{high_dim_infeasible_estimate}
\title{High dimesional infeasible estimator}
\usage{
high_dim_infeasible_estimate(X, y, lambda, maxl2, K, delta = 10,
  verbose = F)
}
\arguments{
\item{X}{\code{n} by \code{d} matrix}

\item{y}{length \code{n} vector}

\item{lambda}{numeric}

\item{maxl2}{numeric}

\item{K}{numeric}

\item{delta}{numeric (minimal spacing between the points to be brute searched over)}

\item{verbose}{boolean}
}
\value{
list containing \code{partition} and \code{coef_list}
}
\description{
High dimesional infeasible estimator
}
