% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/low_dim_estimate.R
\name{low_dim_estimate}
\alias{low_dim_estimate}
\title{Low dimensional estimator}
\usage{
low_dim_estimate(X, y, gamma, delta = 5, verbose = F)
}
\arguments{
\item{X}{\code{n} by \code{d} matrix}

\item{y}{length \code{n} vector}

\item{gamma}{numeric}

\item{delta}{numeric}

\item{verbose}{boolean}
}
\value{
list containing \code{obj_val}, \code{partition}, \code{coef_list}
}
\description{
Low dimensional estimator
}
