% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_dim_buhlmann.R
\name{oracle_tune_gamma}
\alias{oracle_tune_gamma}
\title{Tune gamma (oracle)}
\usage{
oracle_tune_gamma(X, y, lambda, partition, factor = 3/4)
}
\arguments{
\item{X}{\code{n} by \code{d} matrix}

\item{y}{length \code{n} vector}

\item{lambda}{numeric}

\item{partition}{vector with values between 0 and 1}

\item{factor}{numeric}
}
\value{
numeric
}
\description{
Tune gamma (oracle)
}
