% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_dim_buhlmann.R
\name{oracle_tune_gamma_range}
\alias{oracle_tune_gamma_range}
\title{Tune gamma range (oracle)}
\usage{
oracle_tune_gamma_range(X, y, lambda, K, delta = 10, min_gamma = 0.01,
  max_gamma = 16000, max_iter = 10, verbose = T)
}
\arguments{
\item{X}{\code{n} by \code{d} matrix}

\item{y}{length \code{n} vector}

\item{lambda}{numeric}

\item{K}{the number of desired segments (number of changepoints + 1)}

\item{delta}{numeric}

\item{min_gamma}{numeric}

\item{max_gamma}{numeric}

\item{max_iter}{numeric}

\item{verbose}{boolean}
}
\value{
list
}
\description{
Tune gamma range (oracle)
}
