% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_dim_feasible.R
\name{oracle_tune_lambda}
\alias{oracle_tune_lambda}
\title{Tune lambda (oracle)}
\usage{
oracle_tune_lambda(X, y, partition)
}
\arguments{
\item{X}{\code{n} by \code{d} matrix}

\item{y}{length \code{n} vector}

\item{partition}{vector with values between 0 and 1}
}
\value{
numeric
}
\description{
Tune lambda (oracle)
}
