% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screening.R
\name{screening}
\alias{screening}
\title{Screening}
\usage{
screening(fit, tau, M = 100, delta = 1, verbose = F)
}
\arguments{
\item{fit}{object from estimator}

\item{tau}{numeric}

\item{M}{numeric}

\item{delta}{numeric}

\item{verbose}{boolean}
}
\value{
numeric
}
\description{
Screening
}
