% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stationary_timeseries.R
\name{stationary_ar}
\alias{stationary_ar}
\title{Poisson time series, penality is L_{1,1}}
\usage{
stationary_ar(dat, thres_u = round(stats::quantile(dat[dat > 0], probs =
  0.75)), lambda = NA, sparsity = NA,
  basis_function = construct_AR_basis, intercept = T, verbose = F,
  ...)
}
\arguments{
\item{dat}{Count data, where each row represents a different time step and each
column represents a different variable}

\item{thres_u}{A positive threshold for the saturation effect}

\item{lambda}{Tuning parameter for the regression problem, allowed to be \code{NA}}

\item{basis_function}{The function to generate the variables to regress onto, which
can possibly take addition inputs}

\item{intercept}{Boolean to allow for estimating intercepts}

\item{verbose}{Boolean}

\item{...}{Addition parameters for the basis function}
}
\value{
A list containing the final-used \code{lambda} value, the final
objective value (\code{obj_val}), the estimated intercepts (\code{nu}) and
the estimated transition matrix (\code{A}).
}
\description{
Solves the optimization problem (2.8) on https://arxiv.org/pdf/1802.04838.pdf.
The penalty is fixed to be the L_{1,1} (sum of absolute value), but the
construction of the basis can be passed in as a function.
}
\details{
The implementation of the function fits \code{ncol(dat)} Lasso fits under
the Poisson model and concatenates the solutions together.
If lambda is not set, then a heurstical way to concetenate solutions via
cross-validation is used.
}
