% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuning.R
\name{tuning_cross_validation}
\alias{tuning_cross_validation}
\title{Tuning via cross validation}
\usage{
tuning_cross_validation(method, X, y, K_range = c(1:5), cv_verbose = F,
  max_iter = 5, ...)
}
\arguments{
\item{method}{function to fit}

\item{X}{\code{n} by \code{d} matrix}

\item{y}{length \code{n} vector}

\item{K_range}{vector of number of changepoints to consider}

\item{cv_verbose}{boolean}

\item{max_iter}{numeric}

\item{...}{additional parameters for \code{method}}
}
\value{
list
}
\description{
Tuning via cross validation
}
