% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/low_dim_estimate.R
\name{unravel}
\alias{unravel}
\title{Unravel estimate into matrix}
\usage{
unravel(fit)
}
\arguments{
\item{fit}{object from one of the estimation functions}
}
\value{
\code{n} by \code{d} matrix
}
\description{
Unravel estimate into matrix
}
