% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_dim_buhlmann.R
\name{oracle_tune_gamma_hausdorff}
\alias{oracle_tune_gamma_hausdorff}
\title{Tune gamma to minimize hausdorff distance (oracle)}
\usage{
oracle_tune_gamma_hausdorff(X, y, lambda, partition, k_vec, delta = 10,
  min_gamma = 0.01, max_gamma = 1000, verbose = F)
}
\arguments{
\item{X}{\code{n} by \code{d} matrix}

\item{y}{length \code{n} vector}

\item{lambda}{numeric}

\item{k_vec}{numeric vector of the number of desired segments (number of changepoints + 1)}

\item{delta}{numeric}

\item{min_gamma}{numeric}

\item{max_gamma}{numeric}

\item{verbose}{boolean}
}
\value{
numeric
}
\description{
Tune gamma to minimize hausdorff distance (oracle)
}
