% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuning_oracle.R
\name{oracle_tune_grouplambda}
\alias{oracle_tune_grouplambda}
\title{Tune group lambda (oracle)}
\usage{
oracle_tune_grouplambda(X, y, partition, return_refit = F)
}
\arguments{
\item{X}{\code{n} by \code{d} matrix}

\item{y}{length \code{n} vector}

\item{partition}{vector with values between 0 and 1}

\item{return_refit}{boolean}
}
\value{
numeric
}
\description{
Tune group lambda (oracle)
}
